Updates routing logic.
